Feat: 플레이어가 곰을 때릴 수 있는 기능 구현

+++ Assets/02.Scripts/Monster/Bear/BearController.cs
@@ -124,13 +124,13 @@ public class BearController : MonoBehaviourPunCallbacks
         // todo. ApplyVisualState(stateType); 식으로 타 플레이어들에게 애니메이션 호출하기.
     }
 
-    public void RequestDamage(int damage)
+    public void RequestDamage(float damage)
     {
         photonView.RPC(nameof(RPCOnTakeDamage), RpcTarget.MasterClient, damage);
     }
 
     [PunRPC]
-    void RPCOnTakeDamage(int damage)
+    void RPCOnTakeDamage(float damage)
     {
         if (!PhotonNetwork.IsMasterClient) return;
         if (IsInvincible) return;
